pyproject.toml
[tool.poetry]
name = "zillow-bot"
version = "0.1.0"
description = ""
authors = ["Santosh Kalidindi <skalidindi8@gmail.com>"]
readme = "README.md"
packages = [{include = "zillow_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.24.1"
pyairtable = "^1.5.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.4"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.format]
max-line-length = "88"

[tool.poetry.scripts]
dev = "zillow_bot:main"
